Day-10/requirements.txt
streamlit
langchain
langgraph
google-generativeai
pypdf
sentence-transformers
faiss-cpu
